refactor(server): associate "followed" trackers with users
This is in preparation for a "public" tracker feature, where users can
share a Tracker by link and users will select to subscribe to it but the
tracker itself won't be updated with a new "visibleTo" item.

Also changes trackers to no longer update their timestamp when they're
shared; only when they change values.